rename minipupper to mini_pupper
