+++ Python_Module/setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
 name             = MangDang
 version          = 0.1.0
-summary          = MangDang code for minipupper
+summary          = MangDang code for mini_pupper
 description-file = README.rst
 author           = MangDang
 author-email     = fae@mangdang.net
@@ -21,9 +21,9 @@ include_package_data = True
 
 [entry_points]
 console_scripts =
-    calibrate = MangDang.minipupper.calibrate_tool:main
-    cli-calibrate = MangDang.minipupper.calibrate_servos:main
-    set-servos-before-assembly = MangDang.minipupper.set_servos_before_assembly:main
+    calibrate = MangDang.mini_pupper.calibrate_tool:main
+    cli-calibrate = MangDang.mini_pupper.calibrate_servos:main
+    set-servos-before-assembly = MangDang.mini_pupper.set_servos_before_assembly:main
 
 [pbr]
 autodoc_tree_index_modules = True
